+++ sdk/nodejs/apigateway/getResource.ts
@@ -25,11 +25,13 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-aws/blob/master/website/docs/d/api_gateway_resource.html.markdown.
  */
-export function getResource(args: GetResourceArgs, opts?: pulumi.InvokeOptions): Promise<GetResourceResult> {
-    return pulumi.runtime.invoke("aws:apigateway/getResource:getResource", {
+export function getResource(args: GetResourceArgs, opts?: pulumi.InvokeOptions): Promise<GetResourceResult> & GetResourceResult {
+    const promise: Promise<GetResourceResult> = pulumi.runtime.invoke("aws:apigateway/getResource:getResource", {
         "path": args.path,
         "restApiId": args.restApiId,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**