+++ sdk/nodejs/batch/getComputeEnvironment.ts
@@ -21,10 +21,12 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-aws/blob/master/website/docs/d/batch_compute_environment.html.markdown.
  */
-export function getComputeEnvironment(args: GetComputeEnvironmentArgs, opts?: pulumi.InvokeOptions): Promise<GetComputeEnvironmentResult> {
-    return pulumi.runtime.invoke("aws:batch/getComputeEnvironment:getComputeEnvironment", {
+export function getComputeEnvironment(args: GetComputeEnvironmentArgs, opts?: pulumi.InvokeOptions): Promise<GetComputeEnvironmentResult> & GetComputeEnvironmentResult {
+    const promise: Promise<GetComputeEnvironmentResult> = pulumi.runtime.invoke("aws:batch/getComputeEnvironment:getComputeEnvironment", {
         "computeEnvironmentName": args.computeEnvironmentName,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**