+++ sdk/nodejs/batch/getJobQueue.ts
@@ -21,10 +21,12 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-aws/blob/master/website/docs/d/batch_job_queue.html.markdown.
  */
-export function getJobQueue(args: GetJobQueueArgs, opts?: pulumi.InvokeOptions): Promise<GetJobQueueResult> {
-    return pulumi.runtime.invoke("aws:batch/getJobQueue:getJobQueue", {
+export function getJobQueue(args: GetJobQueueArgs, opts?: pulumi.InvokeOptions): Promise<GetJobQueueResult> & GetJobQueueResult {
+    const promise: Promise<GetJobQueueResult> = pulumi.runtime.invoke("aws:batch/getJobQueue:getJobQueue", {
         "name": args.name,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**