+++ sdk/nodejs/ec2/getRouteTable.ts
@@ -35,15 +35,17 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-aws/blob/master/website/docs/d/route_table.html.markdown.
  */
-export function getRouteTable(args?: GetRouteTableArgs, opts?: pulumi.InvokeOptions): Promise<GetRouteTableResult> {
+export function getRouteTable(args?: GetRouteTableArgs, opts?: pulumi.InvokeOptions): Promise<GetRouteTableResult> & GetRouteTableResult {
     args = args || {};
-    return pulumi.runtime.invoke("aws:ec2/getRouteTable:getRouteTable", {
+    const promise: Promise<GetRouteTableResult> = pulumi.runtime.invoke("aws:ec2/getRouteTable:getRouteTable", {
         "filters": args.filters,
         "routeTableId": args.routeTableId,
         "subnetId": args.subnetId,
         "tags": args.tags,
         "vpcId": args.vpcId,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**