+++ sdk/nodejs/ecr/getImage.ts
@@ -21,13 +21,15 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-aws/blob/master/website/docs/d/ecr_image.html.markdown.
  */
-export function getImage(args: GetImageArgs, opts?: pulumi.InvokeOptions): Promise<GetImageResult> {
-    return pulumi.runtime.invoke("aws:ecr/getImage:getImage", {
+export function getImage(args: GetImageArgs, opts?: pulumi.InvokeOptions): Promise<GetImageResult> & GetImageResult {
+    const promise: Promise<GetImageResult> = pulumi.runtime.invoke("aws:ecr/getImage:getImage", {
         "imageDigest": args.imageDigest,
         "imageTag": args.imageTag,
         "registryId": args.registryId,
         "repositoryName": args.repositoryName,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**