+++ sdk/nodejs/getPrefixList.ts
@@ -43,12 +43,14 @@ import * as utilities from "./utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-aws/blob/master/website/docs/d/prefix_list.html.markdown.
  */
-export function getPrefixList(args?: GetPrefixListArgs, opts?: pulumi.InvokeOptions): Promise<GetPrefixListResult> {
+export function getPrefixList(args?: GetPrefixListArgs, opts?: pulumi.InvokeOptions): Promise<GetPrefixListResult> & GetPrefixListResult {
     args = args || {};
-    return pulumi.runtime.invoke("aws:index/getPrefixList:getPrefixList", {
+    const promise: Promise<GetPrefixListResult> = pulumi.runtime.invoke("aws:index/getPrefixList:getPrefixList", {
         "name": args.name,
         "prefixListId": args.prefixListId,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**