+++ sdk/nodejs/rds/getInstance.ts
@@ -20,10 +20,12 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-aws/blob/master/website/docs/d/db_instance.html.markdown.
  */
-export function getInstance(args: GetInstanceArgs, opts?: pulumi.InvokeOptions): Promise<GetInstanceResult> {
-    return pulumi.runtime.invoke("aws:rds/getInstance:getInstance", {
+export function getInstance(args: GetInstanceArgs, opts?: pulumi.InvokeOptions): Promise<GetInstanceResult> & GetInstanceResult {
+    const promise: Promise<GetInstanceResult> = pulumi.runtime.invoke("aws:rds/getInstance:getInstance", {
         "dbInstanceIdentifier": args.dbInstanceIdentifier,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**